add totals, debug

+++ meetyourmapper/views.py
@@ -14,7 +14,9 @@ overpass_map_query = '(node(area:{});<;);out meta;'
 @app.route('/', methods=['get'])
 def index():
 	session['uid'] = str(uuid4())
-	return render_template('index.html')
+	if app.debug:
+		session['osm_file_path'] = 'testdata/testdata.xml'
+	return render_template('index.html', debug=app.debug)
 
 @app.route('/retrieve/<relation_id>', methods=['get'])
 def get_area(relation_id):	
@@ -30,10 +32,11 @@ def get_area(relation_id):
 		'file': session['osm_file_path'],
 		'size': len(resp.content)})
 
-@app.route('/process/<relation_id>', methods=['get'])
-def process_result(relation_id):
+@app.route('/process', methods=['get'])
+def process_result():
 	print(session['osm_file_path'])
 	h = osm.UserHandler()
 	h.apply_file(session['osm_file_path'])
-	os.remove(session['osm_file_path'])
-	return jsonify(h.users)
+	if not app.debug:
+		os.remove(session['osm_file_path'])
+	return jsonify({'totals': h.totals, 'users': h.users})
